Improved Error Handling (better input)

+++ src/io/altar/jseproject/statemachine/states/productstates/ProductRemoveState.java
@@ -4,26 +4,26 @@ import io.altar.jseproject.controller.ProductService;
 import io.altar.jseproject.statemachine.states.State;
 import io.altar.jseproject.statemachine.states.StateType;
 
-import java.util.InputMismatchException;
-import java.util.Scanner;
-
 public class ProductRemoveState implements State {
     @Override
     public StateType executeState() {
-        Scanner input = new Scanner(System.in);
 
-        long id = -1;
-        try {
-            System.out.print("Inserir ID: ");
-            id = input.nextLong();
-        } catch (InputMismatchException e) {
-            //log error
-        }
+        long id = scan("Inserir ID: ", Long.class, false);
 
-        if(ProductService.removeProduct(id) != -1){
-            System.out.println("Produto removido com sucesso!");
+        if (ProductService.getProduct(id) != null) {
+            String confirmation = scan("Deseja remover o produto?\n" +
+                    "[S = remover; Enter = não remover] ", String.class, false);
+            if (confirmation.equalsIgnoreCase("s")) {
+                if (ProductService.removeProduct(id) != -1) {
+                    System.out.println("Produto removido com sucesso!");
+                } else {
+                    System.out.println("Ocurreu um erro! Por favor tente novamente.");
+                }
+            } else {
+                System.out.println("Produto não removido!");
+            }
         } else {
-            System.out.println("Ocurreu um erro! Por favor tente novamente.");
+            System.out.println("Produto Inexistente!");
         }
 
         return StateType.AUTOMATIC_TRANSITION;
