+++ src/io/altar/jseproject/statemachine/states/shelfstates/ShelfRemoveState.java
@@ -4,27 +4,26 @@ import io.altar.jseproject.controller.ShelfService;
 import io.altar.jseproject.statemachine.states.State;
 import io.altar.jseproject.statemachine.states.StateType;
 
-import java.util.InputMismatchException;
-import java.util.Scanner;
-
 public class ShelfRemoveState implements State {
     @Override
     public StateType executeState() {
 
-        Scanner input = new Scanner(System.in);
+        long id = scan("Inserir ID: ", Long.class, false);
 
-        long id = -1;
-        try {
-            System.out.print("Inserir ID: ");
-            id = input.nextLong();
-        } catch (InputMismatchException e) {
-            // log error
-        }
-
-        if(ShelfService.removeShelf(id) != -1){
-            System.out.println("Prateleira removida com sucesso!");
+        if (ShelfService.getShelf(id) != null) {
+            String confirmation = scan("Deseja remover a prateleira?\n" +
+                    "[S = remover; Enter = não remover] ", String.class, false);
+            if (confirmation.equalsIgnoreCase("s")) {
+                if (ShelfService.removeShelf(id) != -1) {
+                    System.out.println("Prateleira removida com sucesso!");
+                } else {
+                    System.out.println("Ocurreu um erro! Por favor tente novamente.");
+                }
+            } else {
+                System.out.println("Prateleira não removida!");
+            }
         } else {
-            System.out.println("Ocurreu um erro! Por favor tente novamente.");
+            System.out.println("Prateleira Inexistente!");
         }
 
         return StateType.AUTOMATIC_TRANSITION;
